+++ jgiven-core/src/main/java/com/tngtech/jgiven/report/text/PlainTextScenarioWriter.java
@@ -146,7 +146,7 @@ public class PlainTextScenarioWriter extends PlainTextWriter {
 
     protected String wordToString( Word word ) {
         if( word.isArg() && !isInt( word ) ) {
-            return "'" + word.getFormattedValue() + "'";
+            return word.getFormattedValue();
         }
         return word.getValue();
     }