+++ jgiven-core/src/test/java/com/tngtech/jgiven/report/text/PlainTextReporterTest.java
@@ -3,6 +3,8 @@ package com.tngtech.jgiven.report.text;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.io.UnsupportedEncodingException;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -13,7 +15,10 @@ import com.tngtech.java.junit.dataprovider.UseDataProvider;
 import com.tngtech.jgiven.GivenTestStep;
 import com.tngtech.jgiven.ThenTestStep;
 import com.tngtech.jgiven.WhenTestStep;
+import com.tngtech.jgiven.annotation.Format;
+import com.tngtech.jgiven.annotation.Quoted;
 import com.tngtech.jgiven.base.ScenarioTestBase;
+import com.tngtech.jgiven.format.BooleanFormatter;
 
 /**
  * Please note that we do explicitly <strong>not</strong> use the ScenarioTest class for JUnit,
@@ -78,7 +83,7 @@ public class PlainTextReporterTest extends ScenarioTestBase<GivenTestStep, WhenT
         GivenTestStep stage = getScenario().addStage( GivenTestStep.class );
         stage.given().a_step_with_a_$_parameter( "test" );
         String string = PlainTextReporter.toString( getScenario().getModel() );
-        assertThat( string ).contains( "Given a step with a 'test' parameter" );
+        assertThat( string ).contains( "Given a step with a test parameter" );
     }
 
     @Test
@@ -87,7 +92,7 @@ public class PlainTextReporterTest extends ScenarioTestBase<GivenTestStep, WhenT
         GivenTestStep stage = getScenario().addStage( GivenTestStep.class );
         stage.a_step_with_a_$_parameter( "test" );
         String string = PlainTextReporter.toString( getScenario().getModel() );
-        assertThat( string ).contains( "a step with a 'test' parameter" );
+        assertThat( string ).contains( "a step with a test parameter" );
     }
 
     @Test
@@ -96,6 +101,38 @@ public class PlainTextReporterTest extends ScenarioTestBase<GivenTestStep, WhenT
         GivenTestStep stage = getScenario().addStage( GivenTestStep.class );
         stage.a_step_with_a_boolean_$_parameter( true );
         String string = PlainTextReporter.toString( getScenario().getModel() );
-        assertThat( string ).contains( "a step with a boolean 'yes' parameter" );
+        assertThat( string ).contains( "a step with a boolean yes parameter" );
     }
+
+    @Format( value = BooleanFormatter.class, args = { "yes", "no" } )
+    @Retention( RetentionPolicy.RUNTIME )
+    @interface YesNo {}
+
+    static class FormattedSteps {
+
+        public void yesno_$_formatted( @YesNo boolean b ) {}
+
+        public void quoted_$_test( @Quoted String s ) {}
+    }
+
+    @Test
+    public void formatter_annotations_are_applied_to_arguments() throws UnsupportedEncodingException {
+        getScenario().startScenario( "test" );
+        FormattedSteps stage = getScenario().addStage( FormattedSteps.class );
+
+        stage.yesno_$_formatted( true );
+        String string = PlainTextReporter.toString( getScenario().getModel() );
+        assertThat( string ).contains( "yesno yes formatted" );
+    }
+
+    @Test
+    public void quoted_is_working() throws UnsupportedEncodingException {
+        getScenario().startScenario( "test" );
+        FormattedSteps stage = getScenario().addStage( FormattedSteps.class );
+
+        stage.quoted_$_test( "foo" );
+        String string = PlainTextReporter.toString( getScenario().getModel() );
+        assertThat( string ).contains( "quoted \"foo\" test" );
+    }
+
 }